Feat: changeRate 타입을 BigDecimal에서 Float로 변경 #101
- StockInfoResponse, StockPriceResponse의 changeRate 필드 타입 변경
- StockPriceService에서 Float 파싱 및 소수점 두 자리 반올림 적용
- API 응답에서 changeRate가 소수점 두 자리로 표시되도록 개선

+++ trading-service/src/main/java/com/example/trading_service/dto/StockPriceResponse.java
@@ -11,7 +11,7 @@ public class StockPriceResponse {
     
     private BigDecimal currentPrice;        // 현재가
     private BigDecimal changePrice;         // 전일 대비 변동가격
-    private BigDecimal changeRate;          // 전일 대비 변동률
+    private Float changeRate;          // 전일 대비 변동률
     private Long volume;                    // 거래량
     private BigDecimal highPrice;           // 고가
     private BigDecimal lowPrice;            // 저가
